backend/requirements.txt
# CPPay Backend - Consolidated Requirements
# Generated from base.txt, development.txt, and production.txt

# ============================================================================
# CORE DJANGO & DRF
# ============================================================================
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
python-decouple>=3.8

# ============================================================================
# DATABASE
# ============================================================================
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# ============================================================================
# CACHING & TASK QUEUE
# ============================================================================
redis>=5.0.0
django-redis>=5.3.0
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# ============================================================================
# API FRAMEWORK (FastAPI)
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ============================================================================
# BLOCKCHAIN & WEB3
# ============================================================================
web3>=6.11.0
eth-account>=0.10.0
eth-utils>=2.3.0
eth-abi>=4.0.0
eth-typing>=3.5.0

# ============================================================================
# HTTP CLIENTS
# ============================================================================
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0

# ============================================================================
# IMAGE PROCESSING (for KYC documents)
# ============================================================================
Pillow>=10.1.0

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.17.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.1.0

# ============================================================================
# CODE QUALITY
# ============================================================================
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.1

# ============================================================================
# DOCUMENTATION
# ============================================================================
mkdocs>=1.5.3
mkdocs-material>=9.5.0

# ============================================================================
# PRODUCTION SERVER & MONITORING
# ============================================================================
gunicorn>=21.2.0
sentry-sdk>=1.38.0
django-silk>=5.0.4

# ============================================================================
# ASGI SERVER
# ============================================================================
daphne>=4.0.0
